# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Moysidis Mobile Massage — practice management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  hasMedicalHistory?: boolean;
}

export interface ClientInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ClientUpdate {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface MedicalHistory {
  id: number;
  clientId: number;
  /** @nullable */
  conditions?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  immuneIssues?: string | null;
  /** @nullable */
  otherHealthIssues?: string | null;
  /** @nullable */
  medications?: string | null;
  /** @nullable */
  painAreas?: string | null;
  previousMassageExperience?: boolean;
  /** @nullable */
  preferredPressure?: string | null;
  /** @nullable */
  massageGoals?: string | null;
  gdprAccepted: boolean;
  /** @nullable */
  gdprAcceptedAt?: string | null;
  createdAt: string;
}

export interface MedicalHistoryInput {
  conditions?: string;
  allergies?: string;
  immuneIssues?: string;
  otherHealthIssues?: string;
  medications?: string;
  painAreas?: string;
  previousMassageExperience?: boolean;
  preferredPressure?: string;
  massageGoals?: string;
  gdprAccepted: boolean;
}

export interface MedicalHistoryUpdate {
  conditions?: string;
  allergies?: string;
  immuneIssues?: string;
  otherHealthIssues?: string;
  medications?: string;
  painAreas?: string;
  previousMassageExperience?: boolean;
  preferredPressure?: string;
  massageGoals?: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export interface Appointment {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  date: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  durationMinutes?: number;
  serviceType: string;
  status: AppointmentStatus;
  price: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  emailScheduledFor?: string | null;
  /** @nullable */
  emailSentAt?: string | null;
  createdAt: string;
}

export interface AppointmentInput {
  clientId: number;
  date: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  durationMinutes?: number;
  serviceType: string;
  price: number;
  /** @nullable */
  notes?: string | null;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export interface AppointmentUpdate {
  date?: string;
  startTime?: string;
  /** @nullable */
  endTime?: string | null;
  durationMinutes?: number;
  serviceType?: string;
  price?: number;
  status?: AppointmentUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface Expense {
  id: number;
  category: string;
  description: string;
  amount: number;
  date: string;
  createdAt: string;
}

export interface ExpenseInput {
  category: string;
  description: string;
  amount: number;
  date: string;
}

export interface ExpenseUpdate {
  category?: string;
  description?: string;
  amount?: number;
  date?: string;
}

export interface FinanceSummary {
  period: string;
  totalIncome: number;
  totalExpenses: number;
  profit: number;
  appointmentCount: number;
}

export interface MonthlyBreakdownItem {
  month: number;
  year: number;
  income: number;
  expenses: number;
  profit: number;
}

export interface ServiceRevenue {
  serviceType: string;
  revenue: number;
  count: number;
}

export interface Review {
  id: number;
  appointmentId: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  emailSentAt?: string | null;
  createdAt: string;
}

export interface ReviewUpdate {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  comment?: string;
}

export interface DashboardStats {
  todayIncome: number;
  todayAppointments: number;
  monthIncome: number;
  monthExpenses: number;
  monthProfit: number;
  totalClients: number;
  avgRating: number;
  pendingAppointments?: number;
}

export type ListClientsParams = {
search?: string;
};

export type ListAppointmentsParams = {
date?: string;
status?: string;
clientId?: number;
};

export type ListExpensesParams = {
month?: string;
year?: number;
};

export type GetFinanceSummaryParams = {
period?: GetFinanceSummaryPeriod;
/**
 * ISO date string — scope for the period
 */
date?: string;
};

export type GetFinanceSummaryPeriod = typeof GetFinanceSummaryPeriod[keyof typeof GetFinanceSummaryPeriod];


export const GetFinanceSummaryPeriod = {
  day: 'day',
  month: 'month',
  year: 'year',
} as const;

export type GetMonthlyBreakdownParams = {
year?: number;
};

export type GetTopServicesParams = {
year?: number;
};