# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Moysidis Mobile Massage — practice management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Register a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a single client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Medical History ───────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get a single appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment details
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments/{appointmentId}/complete:
    post:
      operationId: completeAppointment
      tags: [appointments]
      summary: Mark appointment as completed and schedule thank-you email
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ── Expenses ──────────────────────────────────────────────────────────────
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Add an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{expenseId}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Finance (wow endpoints) ───────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, firstName, lastName, email, phone, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        hasMedicalHistory:
          type: boolean
    ClientInput:
      type: object
      required: [firstName, lastName, email, phone]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ClientUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Medical History ───────────────────────────────────────────────────
...
    Appointment:
      type: object
      required: [id, clientId, date, startTime, serviceType, status, price, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        serviceType:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        price:
          type: number
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        emailScheduledFor:
          type: ["string", "null"]
        emailSentAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [clientId, date, startTime, serviceType, price]
      properties:
        clientId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        serviceType:
          type: string
        price:
          type: number
        notes:
          type: ["string", "null"]
    AppointmentUpdate:
      type: object
      properties:
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        serviceType:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]

    # ── Expense ───────────────────────────────────────────────────────────
    Expense:
      type: object
      required: [id, category, description, amount, date, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, description, amount, date]
      properties:
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string

    # ── Finance ───────────────────────────────────────────────────────────
...