# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Moysidis Mobile Massage — practice management API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  date?: string;
  startTime?: string;
  /** @nullable */
  endTime?: string | null;
  durationMinutes?: number;
  serviceType?: string;
  price?: number;
  status?: AppointmentUpdateStatus;
  /** @nullable */
  notes?: string | null;
}