# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Moysidis Mobile Massage — practice management API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  category?: string;
  description?: string;
  amount?: number;
  date?: string;
}